requirements.txt
Flask>=3.0.0
pandas>=2.0.0
folium>=0.17.0
python-dotenv>=1.0.0
gunicorn>=22.0.0
geopandas>=1.0.0
requests>=2.31.0
